Added a calibration minigame for adjusting audio latency
