+++ Assets/Plugins/FMOD/Cache/Editor/FMODStudioCache.asset
@@ -35,7 +35,7 @@ MonoBehaviour:
   Path: Assets/FMOD Project/Beat-Marker System/Build/Desktop/Master.bank
   Name: Master
   StudioPath: bank:/Master
-  lastModified: 637516744530112658
+  lastModified: 637516802638568515
   LoadResult: 0
   FileSizes: []
   Exists: 1
@@ -106,7 +106,7 @@ MonoBehaviour:
   - {fileID: -6997776824138028954}
   StringsBanks:
   - {fileID: 9016018448889426325}
-  stringsBankWriteTime: 637516744530177654
+  stringsBankWriteTime: 637516802638628527
   cacheVersion: 6
 --- !u!114 &548532804626554394
 MonoBehaviour:
@@ -169,7 +169,7 @@ MonoBehaviour:
   Path: Assets/FMOD Project/Beat-Marker System/Build/Desktop/Master.strings.bank
   Name: Master.strings
   StudioPath: bank:/Master.strings
-  lastModified: 637516744530177654
+  lastModified: 637516802638628527
   LoadResult: 0
   FileSizes:
   - Name: Desktop
